resturcture each packages
