--- packages/synthetics-log-helper/lib/utils.d.ts
@@ -1,4 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-export type OverrideConfiguration = Record<string, string | boolean | string[] | number>;
