+++ packages/synthetics-logger/test-d/index.test-d.ts
@@ -3,10 +3,18 @@
 
 import { expectType } from 'tsd';
 import log = require('../lib');
-import { SyntheticsLogger } from '../lib/SyntheticsLogger';
 
-// Test basic logger export
-expectType<SyntheticsLogger>(log);
+// Test basic logger export - should be an instance
+expectType<typeof log>(log);
 
-// Import other test files
-import './SyntheticsLogger.test-d';
+// Test logging methods
+expectType<void>(log.debug('debug message'));
+expectType<void>(log.debug('debug message', new Error('test')));
+expectType<void>(log.error('error message'));
+expectType<void>(log.error('error message', new Error('test')));
+expectType<void>(log.info('info message'));
+expectType<void>(log.info('info message', new Error('test')));
+expectType<void>(log.log('log message'));
+expectType<void>(log.log('log message', new Error('test')));
+expectType<void>(log.warn('warn message'));
+expectType<void>(log.warn('warn message', new Error('test')));
